+++ packages/button/src/mobile/index.ts
@@ -1,0 +1,2 @@
+export { ButtonMobile } from '../Component.mobile';
+export type { CommonButtonProps as ButtonMobileProps } from '../typings';