+++ packages/input-autocomplete/src/desktop/index.ts
@@ -1,0 +1,1 @@
+export { InputAutocompleteDesktop, InputAutocompleteDesktopProps } from '../Component.desktop';