+++ packages/input-autocomplete/src/mobile/index.ts
@@ -1,0 +1,5 @@
+export { InputAutocompleteMobile, InputAutocompleteMobileProps } from '../Component.mobile';
+export {
+    InputAutocompleteModalMobile,
+    InputAutocompleteModalMobileProps,
+} from '../Component.modal.mobile';