+++ packages/international-phone-input/src/mobile.ts
@@ -1,2 +1,0 @@
-export { InternationalPhoneInputMobile } from './Component.mobile';
-export type { InternationalPhoneInputMobileProps } from './types';