+++ packages/markdown/src/desktop.ts
@@ -1,3 +1,0 @@
-export { MarkdownDesktopComponent as MarkdownDesktop } from './Component.desktop';
-
-export type { MarkdownDesktopProps } from './typings';