+++ packages/navigation-bar/src/shared.ts
@@ -1,2 +1,0 @@
-export * from './components/closer';
-export * from './components/back-arrow-addon';