+++ packages/number-input/src/mobile.ts
@@ -1,1 +1,0 @@
-export { NumberInputMobile, NumberInputMobileProps } from './Component.mobile';