+++ packages/picker-button/src/desktop/index.ts
@@ -1,0 +1,1 @@
+export { PickerButtonDesktop, PickerButtonDesktopProps } from '../Component.desktop';