+++ packages/picker-button/src/shared.ts
@@ -1,2 +1,0 @@
-export { getIcon } from './utils';
-export * from './types';