+++ packages/screenshot-utils/screenshots-story/components.ts
@@ -1,4 +1,4 @@
-const req = require.context('../../', true, /^\.\/(.*)\/src\/(index|desktop|mobile).ts$/);
+const req = require.context('../../', true, /^\.\/(.*)\/src\/(index|(desktop|mobile)\/index).ts$/);
 
 const packages = req.keys().reduce((acc, key) => {
     const packageName = key.split('/')[1];