+++ packages/stepper-input/src/mobile.ts
@@ -1,1 +1,0 @@
-export { StepperInputMobile, StepperInputMobileProps } from './Component.mobile';