+++ packages/stepper-input/src/mobile/index.ts
@@ -1,0 +1,1 @@
+export { StepperInputMobile, StepperInputMobileProps } from '../Component.mobile';