+++ packages/system-message/src/desktop/index.ts
@@ -1,0 +1,2 @@
+export { SystemMessageDesktop } from '../Component.desktop';
+export { SystemMessageDesktopProps } from '../types';