+++ packages/system-message/src/mobile.ts
@@ -1,2 +1,0 @@
-export { SystemMessageMobile } from './Component.mobile';
-export { SystemMessageMobileProps } from './types';