+++ packages/tabs/src/collapsible.ts
@@ -1,2 +1,0 @@
-export { TabsCollapsible, TabsCollapsibleProps } from './components/tabs/Component.collapsible';
-export * from './components/tab';