+++ packages/tooltip/src/mobile/index.ts
@@ -1,0 +1,2 @@
+export { TooltipMobile } from '../Component.mobile';
+export { TooltipMobileProps } from '../types';