+++ tools/rollup/external-with-entry-points.mjs
@@ -1,0 +1,16 @@
+/**
+ *  Добавляем все entryPoints кор-компонентов в externals
+ *  plugin wildcard-external делает это только для js файлов уже после транспиляции ts.
+ *  Если этого не сделать, то плагин rollup-plugin-ts не будет понимать, что зависимости @alfalab/core-components-`*`/shared и др являются внешними и будет дублировать их в каждом пакете
+ */
+export default function (depsArr) {
+    return depsArr.reduce((acc, dep) => {
+        if (dep.startsWith('@alfalab/core-components-')) {
+            acc.push(dep, dep + '/mobile', dep + '/desktop', dep + '/shared');
+        } else {
+            acc.push(dep);
+        }
+
+        return acc;
+    }, []);
+}