+++ packages/input-autocomplete/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 10.3.0
+
+### Minor Changes
+
+### [#944](https://github.com/core-ds/core-components/pull/944)
+
+-   Добавлен package.json с module полем в mobile, desktop, shared точки входа
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@9.1.0
+    -   form-control@10.2.0
+    -   input@12.3.0
+    -   select@15.3.0
+    -   shared@0.7.0
+
 ## 10.2.3
 
 ### Patch Changes