+++ packages/pagination/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.1.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   tag@6.1.0
+
 ## 2.1.2
 
 ### Patch Changes