+++ packages/toast-plate/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 6.1.0
+
+### Minor Changes
+
+### [#944](https://github.com/core-ds/core-components/pull/944)
+
+-   Добавлен package.json с module полем в mobile, desktop, shared точки входа
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.3.2
+
 ## 6.0.7
 
 ### Patch Changes